Se agrego actualizacion images

+++ src/main/java/com/example/demo/controllers/dto/ProductDTO.java
@@ -1,8 +1,5 @@
 package com.example.demo.controllers.dto;
 
-import java.math.BigDecimal;
-
-
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -17,7 +14,7 @@ public class ProductDTO {
 
     private Long id;
     private String name;
-    private BigDecimal price;
+    private int price;
     private String category;
     private String images;
